[1.0.6] Handle exception when polling for work so addon doesn't die
In the event there is connectivity problems from the HA host to
AWS, the calls to AWS result in exceptions that kill the work. Even
with watchdog enabled the addon still dies because the addon tries
to call AWS again as soon as it starts up. This commit handles
errors when polling for work which is the most likely scenario for
errors to happen. Its still possible for error to happen during
calling CloudFormation (statup) or sending the response back
to SQS.

https://github.com/adamirr/ha_addons/issues/2